Worker build using custom webpack config.
