+++ .eslintignore
@@ -1,0 +1,1 @@
+datatables.bootstrap4.js
