+++ angular.json
@@ -3,6 +3,18 @@
   "version": 1,
   "newProjectRoot": "projects",
   "projects": {
+    "worker-app": {
+      "root": "",
+      "projectType": "library",
+      "architect": {
+        "build": {
+          "builder": "@angular-devkit/build-webpack:webpack",
+          "options": {
+            "webpackConfig": "worker.webpack.config.js"
+          }
+        }
+      }
+    },
     "angular-app": {
       "root": "",
       "sourceRoot": "src",
@@ -27,7 +39,8 @@
               "node_modules/jquery/dist/jquery.min.js",
               "node_modules/datatables.net/js/jquery.dataTables.js",
               "src/app/widgets/datatables/datatables.bootstrap4.js",
-              "node_modules/chosen-js/chosen.jquery.min.js"
+              "node_modules/chosen-js/chosen.jquery.min.js",
+              {"input": "dist/decoder-worker.js", "lazy": true }
             ]
           },
           "configurations": {
