--- src/app/pages/apps/pages/audio/components/audioevents/worker.js
@@ -1,6 +1,0 @@
-self.onmessage = event => {
-  if (event.data.decode) {
-    console.log('DATA RECEIVED');
-    //self.postMessage(decoded, decoded.channelData);
-  }
-}
